WIP move last 4 problematic tests (these will all fail)
